Changed Integer to Double within each hero Class

+++ app/src/main/java/com/planarform/daniel/owstats/OWAna.java
@@ -7,23 +7,23 @@ import java.util.Map;
  */
 
 public class OWAna extends OWHero{
-    Integer scopedShots;
-    Integer nanoBoostAssistsMostInGame;
-    Integer unscopedAccuracy;
-    Integer scopedHits;
-    Integer nanoBoostsApplied;
-    Integer scopedAccuracyBestInGame;
-    Integer nanoBoostAssists;
+    Double scopedShots;
+    Double nanoBoostAssistsMostInGame;
+    Double unscopedAccuracy;
+    Double scopedHits;
+    Double nanoBoostsApplied;
+    Double scopedAccuracyBestInGame;
+    Double nanoBoostAssists;
 
     public OWAna(Map stats) {
         super(stats);
-        this.scopedShots = (Integer)stats.get("scoped_shots");
-        this.nanoBoostAssists = (Integer)stats.get("nano_boost_assists_most_in_game");
-        this.unscopedAccuracy = (Integer)stats.get("unscoped_accuracy");
-        this.scopedHits = (Integer)stats.get("scoped_hits");
-        this.nanoBoostsApplied = (Integer)stats.get("nano_boosts_applied");
-        this.scopedAccuracyBestInGame = (Integer)stats.get("scoped_accuracy_best_in_game");
-        this.nanoBoostAssists = (Integer)stats.get("nano_boosts_assists");
+        this.scopedShots = (Double)stats.get("scoped_shots");
+        this.nanoBoostAssists = (Double)stats.get("nano_boost_assists_most_in_game");
+        this.unscopedAccuracy = (Double)stats.get("unscoped_accuracy");
+        this.scopedHits = (Double)stats.get("scoped_hits");
+        this.nanoBoostsApplied = (Double)stats.get("nano_boosts_applied");
+        this.scopedAccuracyBestInGame = (Double)stats.get("scoped_accuracy_best_in_game");
+        this.nanoBoostAssists = (Double)stats.get("nano_boosts_assists");
 
 
     }
