+++ app/src/main/java/com/planarform/daniel/owstats/OWPharah.java
@@ -7,18 +7,18 @@ import java.util.Map;
  */
 
 public class OWPharah extends OWHero {
-    Integer barrageKills;
-    Integer barrageKillsMostInGame;
-    Integer meleeFinalBlowMostInGame;
-    Integer rocketDirectHitsMostInGame;
-    Integer rocketDirectHits;
+    Double barrageKills;
+    Double barrageKillsMostInGame;
+    Double meleeFinalBlowMostInGame;
+    Double rocketDirectHitsMostInGame;
+    Double rocketDirectHits;
 
     public OWPharah(Map stats){
         super(stats);
-        this.barrageKills = (Integer)stats.get("barrage_kills");
-        this.barrageKillsMostInGame = (Integer)stats.get("barrage_kills_most_in_game");
-        this.meleeFinalBlowMostInGame = (Integer)stats.get("melee_final_blows_most_in_game");
-        this.rocketDirectHitsMostInGame = (Integer)stats.get("rocket_direct_hits_most_in_game");
-        this.rocketDirectHits = (Integer)stats.get("rocket_direct_hits");
+        this.barrageKills = (Double)stats.get("barrage_kills");
+        this.barrageKillsMostInGame = (Double)stats.get("barrage_kills_most_in_game");
+        this.meleeFinalBlowMostInGame = (Double)stats.get("melee_final_blows_most_in_game");
+        this.rocketDirectHitsMostInGame = (Double)stats.get("rocket_direct_hits_most_in_game");
+        this.rocketDirectHits = (Double)stats.get("rocket_direct_hits");
     }
 }
