+++ app/src/main/java/com/planarform/daniel/owstats/OWSoldier76.java
@@ -7,23 +7,23 @@ import java.util.Map;
  */
 
 public class OWSoldier76 extends OWHero {
-    Integer helixRocketsKillsMostInGame;
-    Integer tacticalVisorKills;
-    Integer helixRocketsKills;
-    Integer meleeFinalBlowsMostInGame;
-    Integer tacticalVisorKillsMostInGame;
-    Integer bioticFieldsDeployed;
-    Integer bioticFieldHealingDone;
+    Double helixRocketsKillsMostInGame;
+    Double tacticalVisorKills;
+    Double helixRocketsKills;
+    Double meleeFinalBlowsMostInGame;
+    Double tacticalVisorKillsMostInGame;
+    Double bioticFieldsDeployed;
+    Double bioticFieldHealingDone;
 
     public OWSoldier76 (Map stats) {
         super(stats);
-        this.helixRocketsKillsMostInGame = (Integer)stats.get("helix_rockets_kills_most_in_game");
-        this.tacticalVisorKills = (Integer)stats.get("tactical_visor_kills");
-        this.helixRocketsKills = (Integer)stats.get("helix_rockets_kills");
-        this.meleeFinalBlowsMostInGame = (Integer)stats.get("melee_final_blows_most_in_game");
-        this.tacticalVisorKillsMostInGame = (Integer)stats.get("tactical_visor_kills_most_in_game");
-        this.bioticFieldHealingDone = (Integer)stats.get("biotic_field_healing_done");
-        this.bioticFieldsDeployed = (Integer)stats.get("biotic_fields_deployed");
+        this.helixRocketsKillsMostInGame = (Double)stats.get("helix_rockets_kills_most_in_game");
+        this.tacticalVisorKills = (Double)stats.get("tactical_visor_kills");
+        this.helixRocketsKills = (Double)stats.get("helix_rockets_kills");
+        this.meleeFinalBlowsMostInGame = (Double)stats.get("melee_final_blows_most_in_game");
+        this.tacticalVisorKillsMostInGame = (Double)stats.get("tactical_visor_kills_most_in_game");
+        this.bioticFieldHealingDone = (Double)stats.get("biotic_field_healing_done");
+        this.bioticFieldsDeployed = (Double)stats.get("biotic_fields_deployed");
     }
 
 }
